requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
filelock==3.13.3
fsspec==2024.3.1
future==1.0.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai-whisper==20231117
pillow==10.3.0
proglog==0.1.10
pyee==11.1.0
python-ffmpeg==2.0.10
pytube==15.0.0
regex==2023.12.25
requests==2.31.0
sympy==1.12
tiktoken==0.6.0
torch==2.2.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
